警告をエラーにして、色々直した

+++ 3DDirectX1/others/CollisionManager.h
@@ -37,7 +37,7 @@ public:// ƒƒ“ƒoŠÖ”
 	float GetTsize() { return Tsize3; }
 	float SetTsize2(float Tsize2) { return this->Tsize2 = Tsize2; }
 	float SetTsize(float Tsize) { return this->Tsize = Tsize; }
-	float GetHitCount() { return HitCount; }
+	int GetHitCount() { return HitCount; }
 	bool GetHit() { return IsHit; }
 	bool SetParentFlag(bool parentFlag) { return this->PFlag = parentFlag; }
 private:
